Majority Classes are Working
point, polygon and node are working. Currently. They are polished up and should not otherwise need to be changed. 

Some changes were made to node though which was working for previous assignments which is concerning. 

In order to assign node data I had to make polygon a constructor with no arguments and I don't understand why that fixed it. 

I am concerned about the use of const (I am currently not using it at all) and whether or not I have to use it in my assignment. If I do have to use it I am worried I will then need to make a large number of fundamental changes to the work I have done.

Finally I have started working on MyPolygons (the linked list). Very little has been done asside from create the header shell.
Peace Out